error sequence table
