+++ screen2.0/src/app/applets/argo/types.ts
@@ -287,15 +287,19 @@ export type TomtomMatchQueryData = {
 export type SequenceTableProps = {
     sequenceFilterVariables: SequenceFilterState;
     SubTableTitle: React.FC<SubTableTitleProps>;
-    sequenceRows: SequenceTableRow[];
+    inputRegions: InputRegions;
     isolatedRows: IsolatedRow;
+    updateSequenceRows: (rows: SequenceTableRow[]) => void;
+    updateLoadingSequenceRows: (loading: boolean) => void;
 }
 
 export type ElementTableProps = {
     elementFilterVariables: ElementFilterState;
     SubTableTitle: React.FC<SubTableTitleProps>;
-    elementRows: ElementTableRow[];
+    intersectingCcres: CCREs;
     isolatedRows: IsolatedRow;
+    updateElementRows: (rows: ElementTableRow[]) => void;
+    updateLoadingElementRows: (loading: boolean) => void;
 }
 
 export type GeneTableProps = {
